Adjusted HUD, Tested some levels

+++ Assets/Entities/Payload/PayloadBehaviour.cs
@@ -95,7 +95,7 @@ public class PayloadBehaviour : Entity
         //Debug.Log("Starting to fill gas");
         while (fillingGas)
         {
-            HUDController.Instance.SetPayloadGas(currentGas / maxGas);
+            HUDController.Instance.SetPayloadGas((float)currentGas / (float)maxGas);
             count += Time.deltaTime;
             if (count > 1f / fillingRate) 
             {
@@ -148,7 +148,7 @@ public class PayloadBehaviour : Entity
         {
             PayloadBehaviour.Instance.agent.isStopped = false;
             //Debug.Log($"Payload Moving, current Gas {currentGas}");
-            HUDController.Instance.SetPayloadGas(currentGas / maxGas);
+            HUDController.Instance.SetPayloadGas((float)currentGas / (float) maxGas);
             count += Time.deltaTime;
             if (count > 1f / burningRate)
             {
